Added -r8 flag to tapenade. We will compile flow and ajoint routines in double precision. Hence it is necessary to generate double precision differentiated routines. While compiling these differentiated routines, use -r8 or equivalent flag for your compiler. This is set in the top makefile.in file through FFLAGS variable.
QB is set to zero in main routine and not in resid_q.F. The position of BC_Q in main.F was changed in version 28; this is wrong and has been changed back. 

SB was set to zero in resid_x.F which is wrong and has been removed.


git-svn-id: https://svn.berlios.de/svnroot/repos/nuwtun/trunk@31 b4dca852-493c-0410-a6d7-ee0e29af0bb2